[bt][bt-gap] Don't wait for OnAdaptersFound for get_adapters
The OnAdaptersFound mechanism bt-gap uses to provide higher likelihood
of an adapter being available for inquiries is not relevant for the
get_adapters() call which should just return the current state of the
available hosts. This is also motivated by removing the 5 second wait
time from the integration tests, which currently begin by calling
get_adapters() to ignore previously available hosts.

test: compiles; functionality has not changed apart from the wait.

Change-Id: Ib4676026d918f7da4de8555e87ab78f03cf826b5