[scenic] Move Present()-callbacks to after actual presentation.
Currently the callbacks for calls to Present() trigger after the updates are applied, but before they're actually presented, giving an estimated presentation time as a result.
This also has the side effect of triggering rendering of the next frame in apps that listen to that callback, which causes lag due to contention with Scenic at the most crucial time.

This CL moves the callback to after content actually reaches the screen, returning the real presentation time, solving both of these issues.

Also: Fixed two bugs related to rescheduling.
1. Updates weren't being rescheduled if their fences hadn't been met.
2. If an update didn't make any changes it sometimes ended up cancelling the scheduling of later updates.

Change-Id: I197c773763a3658a02db43478d931ca904733364