[rust][scripts] Fix Rust fx scripts
These scripts rely on the location of the build directory.
The process they used to determine the location of this directory
was invalidated as a result of `fx` changes. In order to fix this
and prevent further breakages, these scripts are now called from
a thin shell wrapper which provides the necessary environment
variables by sourcing ./vars.sh.

Change-Id: Ia257a5349042da1cd5e88af4186d815ab7650d71