[build] Use buildidtool to manage .build-id
This change shows how .build-id can be managed with a special tool I've created

Stamp files act as unique pointers to .build-id entries and are thus almost
first class citizens of the build. For non-C/C++ builds these stamp files
are taken as order only deps. This means that they're treated as if they're
useful at runtime which is perfect since they're stand ins for the .build-id
entries. In the C/C++ build these are run at every link step which isn't ideal
for paralellism reasons but it still ensures that if the binary exists then
the stamp file and thus the .build-id entry exists which is good enough.
Once we have multiple action toolsets and native GN support for rust and go
we can make all of that function in an ideal way.

Test: built Fuchsia every way I could + CQ
Change-Id: I97ba2c9f12c574a06cd5539dc9cf175fb11108d1