[build][fuzzing] Use components in GN templates
This CL is a step towards fixing fuzzing on Fuchsia.  It updates the
fuzz_target and fuzz_package templates to use the component template. It
also drops the superfluous corpora variable, as we've subsequently
landed on CIPD, which can do the version tracking more naturally.
Finally, it disables the zircon fuzzers until ZX-3624 is complete.

Note that this CL is not fully compatible with the (currently broken)
'fx fuzz' command.  Follow-on CLs will address those issues.

SEC-225

Test: fx set core.x64 --with //bundle:tests --fuzz-with asan

Change-Id: I3f1e062cfd1044a55a16f80f9808e37d269dbe1d