[zircon][gn] Add barriers to zbi() metadata collection
If zbi("A") depends on zbi("B"), A uses B as a zbi_input() but it doesn't
see into B's deps and redundantly collect those zbi_input() items.

If a resource() depends on a zbi(), then that zbi() doesn't act as a
zbi_input() to any zbi() that depends on the resource().

Before this, the zbi-bootfs test-image.zbi gets folded into the main
BOOTFS image because it's included as a zbi_input() inadvertently by
dint of being in deps of a resource().

Change-Id: I668267d2c76c9ced547fdc1c1e071444c644c76f