[kernel][thread] try to detect a corrupted idle thread at bootup
ZX-3672 concerns what appears to be a partially intialized secondary
idle thread at boot time when coming out of a mexec. The current working
theory is a secondary cpu from the previous kernel boot is not getting
its cache lines completely flushed until after the new kernel instance
has been copied in place.

To try to catch this, hex dump the area around the thread to see if it's
been partially initialized from the previous system state when it's not
supposed to.

Tested: boot and run on x86 and arm, make sure it doesn't falsely
trigger.

Change-Id: I00d8a9617c0796dd1cb081a1d209e8310b8373f9