[kernel][threads] Add NO_BOOST flag to threads
Add a flag to thread which can be used to suppress the kernel's
"priority boost" behavior.  Make sure that the scheduler boost
behavior obeys this flag.

Note: there is *no* API for this functionality.

The only reason to do this at all is to make it practical to write
kernel unit tests which need to validate the priority inheritance
chaining behavior.  In order to do this, the tests need to be able to
deterministically control the base priority of the test threads it
creates so that they can create arbitrary inheritance chains and
validate the expected resulting effective priority of the threads
involved in the test.  Allowing these threads to be subject to boost
behavior effectively makes this base priority non-deterministic, which
is why it needs to be suppressed.

No normal code should ever be using this flag.  It is for specific
testing purposes only.

Change-Id: Ic02fc9851cda28ff15250a6a6d9b46a52c5a06f7