[e2e] Dump captured audio too.
Like we dump captured screenshots, we also dump captured audio.
Reuses the Dump class factored out of Scenic. Of course we would
like to reuse the instance not just the class, but that requires
an API change which would be breaking integration, hence it
will be separate.

TESTED=ran test, checked dumped wav file on https://cloud.google.com/speech-to-text/

Change-Id: I3fd04f6a75c393649fe16e0035a64b953495f501