[pkg-cache] Update fidl to the latest v2 design
The v2 design for the package cache has been changed to replace
the `BlobFetcher` interface with a `NeededBlobs` iterator
interface. This updates the fidl to reflect this change, and
updates the package cache to compile with the rewrite.
The `PackageCache::Get` method is not actually used yet, so the
implementation is stubbed out.

Test: Verified Fuchsia fully compiles.
Bug: PKG-599 #done
Change-Id: Ia064ccd3b62ad047078441d6b1f2e396f9a4e110