[fidl][cpp] fidl::Equals for all FIDL types
A fidl::Equals function template tests two const references of the same
type for deep equality. This is implemented as a call to
fidl::Equality<T>::Equals(...) which is generated for every generated
FIDL type and implemented for std::vector<T>, std::string, etc.

To allow a soft transition the old operator== declarations remain
guarded by a FIDL_OPERATOR_EQUALS macro that is currently defined. This
CL fixes up all of fuchsia.git to not need that defined.

operator== is defined on fidl::Array of integer types in anticipation of
replacing fidl::Array with std::array.

FIDL-563

Change-Id: Iec2307b641305f3f46e8d2205bd41c1bcdf2de55