[zircon][hid] Create HID driver integration test
Create an integration test for the HID driver
using Hidctl. Hidctl spoofs the Hidbus protocol
and allows us to send fake report descriptors
and reports.

The current test sends a Mouse report descriptor
and verifies that the correct /dev/class/input/
file is created, and that reports can be recieved
with the correct size and information.

ZX-3813 #done

Change-Id: I0bcf5d43ec89cd6a303b9a3feeee5a10d8fd79fe