[debugger] Implicit casting of derived collections.
If there is a Base class and a Derived class, this patch allows
assigning Derived to Base, and converting Derived* to Base* with proper
pointer adjustments.

This currently only handles the cases required for implicit casting.
Explicit static_casting will be able to convert in the opposite
direction and can cast references which are not handled/tested in this
patch.

Update VisitClassHierarchy to take a uint64_t which is how derived type
offsets are stored.

Change-Id: Ifb25f87c8f091fb9619b094d8cd34da927b0c651