[zxcrypt] more xprintfs, less chained assignments
Sharing "propagate failure" return impls are great right up until one of
the conditions is failing and you can't tell which one because you can't
use a debugger and you need to tell which statements are succeeding and
which is returning an error.

Since xprintf is removed at compile-time unless ZXDEBUG is defined
nonzero, this should compile into the same code in practice, but as soon
as something is going awry we can see specifically what in one
edit/compile/test cycle.

Change-Id: Ibb5fcb215c9192c2b206dc18e5bca77b80e394ad