[handle][musl] Update handle bitfield packing.
Update the way in which bits are allocated in handles, and formalize
the guarantees we provide moving forward.

Prior to this, user-mode encodings of handles were 32 bits, and had
the following guarantees.
1) The value 0 always indicated an invalid handle.
2) The LSB of the handle would always be 1.
3) The MSB of the handle would always be 0.

Unfortunately, #1 was the only documented guarantee.  #2 and #3 were
undocumented, but #3 had been implied a very long time ago when we
used to use non-negative values as handle return types, and negative
values as error codes.  The practice of mixing zx_handle_t and
zx_status_t ended a long time ago, however.

Still, these guarantees are important for some code; specifically for
things like futex based mutex implementations which take advantage of
the spare bits in order to record mutex ownship as well as contested
state in their atomic futex storage.

Now, the guarantees have been updated and are...
1) The value 0 will indicate an invalid handle.
2) The two LSBs of the handle will always be 1.

Note: Because of the nature of the kernel -> user mode handle
obfuscation, it is not possible to guarantee that a user mode handle
will have any of its top 30 MSBs set.  In particular, the value
0x00000003 can absolutely be a valid user-mode handle encoding.

These guarantees have now been formally stated in the zircon
documentation which discusses zx_handle_t.

Also, update MUSL's pthread_mutex, R/W lock, and condvar to align with
the new packing.  Prior to this, there were a large number of places
where the pthread mutex code made assumptions about the structure of
the handle bits.  Assumptions were both implicit (such as comparing
(state < 0) to determine if a handle/contested state was contested)
and explicit (such as or'ing the literal 0x80000000 into a handle in
order to encoded an owned-and-contested mutex state).

So; this CL also makes an attempt to clean this up a bit.  It
centralizes all of the state manipulations in static inline functions,
and removes all of the explicit/implicit assumptions about state.
Finally, it aligns all of the state manipulations with the current
definition of handle bits outlined above.

Tests: runtests on x64/arm64 QEMU, Intel NUC and VIM2

Change-Id: I80eba73b3783d3dd1236261ca6e962424a7bcf6c