[linux_runner] Avoid calling ftruncate on existing file.
This was appearing to take 40s in traces of the VM startup time,
however removal of this call does not have an appreciable difference on
VM startup time. What I suspect is happening is the |open| is causing
the filesystem to begin reading the file. The |ftruncate| is then a
serializing operation.

Removal of the |ftruncate| doesn't actually speed up the filesystem,
but it allows the VM to begin booting until it actually needs to read
off the disk.

Even though this won't speed up startup times, calling ftruncate on an
existing file can be a bit dubious, especially since we won't resize
any filesystems on the disk anyways. Its removal also makes the traces
easier to read.

Test: scp'd an existing stateful.img, verified boot was successful.
      Removed the stateful.img and verified a new, 4GB file was created
      on the next boot.
Change-Id: I70b175509079214d09d6154d80115c4299b72a7e