[bt][sdp] Add retrieve-all option for discovery
While possible to specify all attributes before, the API was unwieldy
and required you to pass around objects with 65535 objects.  In this
common case, allow all attributes to be requested by requesting no
specific attributes.

This is mostly updating documentation throughout, and adding the support
for the behavior in SDP Client.

Test: GAP_BrEdrConnectionManager.AddServiceSearchAll
Change-Id: Iecbeb247779cdd0d524c64e0dff942ffe472856f