[scripts][fuzzing] Use GN metadata `fx fuzz list`
This CL refactors how the list of fuzzers is determined.  Instead of
scanning all packages installed on a device, we emit metadata from the
GN templates in the form of a fuzzers.json file, and read it using a
python script.

The python in this CL is a bit more complex than what's strictly
necessary for the task of listing fuzzers.  It introduces some of the
objects and patterns that will be used to reimplement more of 'fx fuzz'
in subsequent CLs.

SEC-225

Test: python scripts/fuzzing/test/fuzzer_test.py
Change-Id: I17637d5b187afff51478b115df322f17409c7122