[virtio][scsi] Increase SCSI max xfer size.
Query the SCSI target's BLOCK LIMITS VPD page (if present) and set the
max xfer size to the lower of what that page exports and the virtio
controller's config knob (with a hard limit of 512KB). Going from a 16K
xfer size to a 256K xfer size more than doubles thruput on the QEMU.
The gains are similar on GCE (but much more variable).

Bug: ZX-3753

Test: iochk and biotime. Both on GCE and QEMU.
Change-Id: I2205d19b15dc0cd9e1a23951bda8d119ef978ef6