[debugger] Add static cast.
Implement a static cast which can convert pointers and references to
base classes to ones to derived classes. This is the only new
functionality over the existing implicit casting because the debugger's
implicit cast is so powerful.

Adds tests for the new static_cast, C casts, and some void* implicit
casting cases that were missted before.

Split out the elaborate derived type setup used in the existing derived
type casting test to a shared header where it can be used by other
tests.

Change-Id: I6b78e72012491e39ee5e8599c5879f8bb83572d1