[fidl][cpp] Put generated files in a per-target dir
We've had weak dependency enforcement between generated C++ bindings and
the code that uses them. This has been the source of some build flakes.
Nobody likes build flakes.

With this change generated C and C++ bindings for FIDL libraries are
placed under a directory named after the GN fidl() target that declared
them. This means that to get to the headers the build file for the
referencing code must know (possibly transitively) which target defined
the library.

FIDL-535 #done

Change-Id: Ic0d637862e0b4d95006d4e88db280ea3fef900ab