[sequential_futures] Add sequential_futures library
This library provides functionality to execute Futures sequentially.

A SequentialSender is used to schedule a future for execution once all
the previously sent futures have completed.

SequentialSender::new() returns a new sender as well as a Future which
drives the execution of the sent futures.

Note: This is similar to futures::stream::FuturesOrdered except
FuturesOrdered allows the execution of the futures to race and only
guarantees the results to be yielded in order.

Change-Id: I30d6d8fc3751e9296f5bc0e66941dd83aa1dbe1f