[paver] Move paver backend to a separate lib.
This is the first change in making paver into a component service, which
exposes a FIDL interface. The second step will be to take create a
service lib which implements the FIDL interface as a frontend to the
same backend library. The final change will be to take the paver binary
and utilize the FIDL service rather than talk directly to the backend.

ZX-3861

Change-Id: Iba916fb0c9733ace0257b8fe4e73b86f909447f0