[debugger] Separate out name finding context.
There are a lot of parameters (many optional) to FindName and variants
which made them ungainly to call. A planned addition for prefix
searching will add even more to search only the current module in a
running proces or all modules in a non-running process in a non-running
process (currently neither of these are expressable).

This patch separates out all of the context rqeuired by find name. This
eliminates some boilerplate and clarifies what is optional and what
exactly different types of context mean. It adds the ability to express
the types of context mentioned above.

Change-Id: Ie5e41dec67f4c68a5280910e2d4328debcee010b