[text] Add specific index conversion functions
It turns out Flutter expects selection and composition range indices to
be specified in UTF-16 code units. Since Rust strings are UTF-8, it
seems impractical to have the unit throughout the code be these code
units, not to mention the existing code already expects bytes. This
change adds a conversion step any time a legacy TextInputState is either
sent in from a InputMethodEditorClient, or we send a DidUpdateState to a
client, which allows the clients to speak these UTF-16 code units,
and us to internally continue to use bytes.

Optimally, we'd refactor out the byte usage and only use code points,
but this seems like overkill right now considering many parts of ime.rs
will be deleted after the migration to IME API v2.

This change also refactors out some byte index -> code point index
conversion that was happening in the proxy to the new TextField
interface, which uses code points.

(To make the distinction clear — a code unit is a two-byte chunk in
UTF-16, analogous to bytes in UTF-8. One or two code units in UTF-16, or
one to four bytes in UTF-8, comprise a Unicode code *point*.)

A couple of tests in ime.rs have been changed to reflect the index unit
change. I've also modified one test for an out-of-range selection that I
believe had incorrect behavior.

Change-Id: I11cc6689658952f5d0a43e5cceb4ad6ab58d17a0