[httpdate-hyper] Add an HTTPS fallback for network time
A simple library providing a single API to retrieve a the current time
over an HTTPS connection.

Bug: CF-701

This requires some contortions during trust evaluation, because evaluation
requires the current time:

* First we blindly trust the certificates we receive during the handshake,
  and then parse out date from the http date header.
* We then verify the date is at least slightly plausible by requiring it
  be after the system build date.
* Finally we perform a true trust evaluation against the certificates we
  received during the handshake, using the response date.

Obviously this is less than ideal, but it is the best we can do given the
constraint that we don't already know the current date and time. By
validating against the system rootstore, and verifying that the response
date is after the build date, we've at least blocked expired certs, and
requiring that the server cert was issued by a real CA.

There is no revocation checking, as it is unclear if that would have true
value given the limited trust we have in the entire connection.

In order to achieve the deferred certificate validation it was necessary
for us to opt in to the "dangerous_configuration" feature of rustls.

Our tests do hit the network, but consider network failures to be valid
results.

Change-Id: If71f4153d3c7b7bc5c3f1f01749093ecbbfeccc3