[mdns] change FIDL definition for subscribers
This CL changes the way mdns clients get an up-to-date collection of
service instances when the call SubscribeToService. Instead of
passing a request for a ServiceSubscription that has an event, the
client passes a handle for a ServiceSubscriber that has methods.
The methods (basically add/change/remove) have replies, allowing
the service to throttle traffic.

Also, a 'perform_probe' parameter has been added to the two publish
methods, allowing the client to suppress a probe for conflicting
instances. This speeds things up when instance names are assumed to
be unique (e.g contain a GUID).

All subscribing clients have been updated to use the new definition.

The mdns cpp client library has been deleted, because the new
approach is easier for clients to handle without that support. The
old approach delivered the entire collection of instances every
time there was a change, and the library notified the client of
what changes had occurred. The new approach sends changes, which,
in most cases, is what the client wants anyway.

TEST: no behavior change
      $ fx shell mdns-util subscribe _fuchsia._udp.

NET-1842 #done

Change-Id: I32f001cc722f78738f33dfb350797fd09761d990