[text] Replace normal Mutex with futures-aware Mutex in hardware IME
Fixes a race condition where if an OnUpdate came in while a the state
lock was held (due to a inflight request), it would request the same
lock a second time and deadlock, since the thread would be blocked and
never return back to the inflight request. By using Mutexs that allow
await!ing, we can allow the executor to continue work on the inflight
request even when someone else is requesting a lock.

Change-Id: Id745bff801f72e35ee09bd219b78df769a4b2776