[virtio][scsi] Eliminate a copy in the virtio scsi data path.
Eliminate one of the 2 copies in the RW path, by mapping the vmo and
doing IO in/out of that mapped address. There is still a copy in/out
of the iobuffer in the scsi code. The plan is to remove that later by
pinning the pages in the vmo and setting up the DMA by iterating over
the paddrs.

Bug: ZX-3754.

Test: iochk and biotime on QEMU. Boot, run biotime on GCE.
Change-Id: Icacf13a14f3c5c155463e60cae274860d20da940