[mediaplayer] discard no-pts packet during audio warmup
This CL causes the audio renderer wrapper in in mediaplayer to
discard packets if:

1) the audio renderer itself has not sent any messages
2) the packets in question have no timestamps

This addresses a problem encountered in bluetooth A2DP in which
packets are pushed regardless of demand. If the packages
implementing the audio renderer need to be loaded, the incoming
packets can overflow the VMO used to send content to the renderer,
causing mediaplayer to panic. With this CL, mediaplayer discards
packets until the arrival of an OnMinLeadTimeChanged event, which
is our first indication that the renderer is actually running.

TEST: manually tested by streaming from pixelbook to pixel 3

MTWN-252 #done

Change-Id: Ie735d65ad73e1e5143a87fb18ad8949c6189ab55