[pseudo-fs] add handle_close logic to connection object.
this moves the logic relating to the connection specific state during
the close operation to the connection object directly. the function
takes a closure for writing the actual data. the split in logic needs to
continue to happen because the connection doesn't have access to
on_write, and in the later async version of the pseudo file, the
on_write_futures list, so there is still pseudo file specific work to be
done, but this lets us hide the internals of the connection object more
effectively.

Change-Id: I2dec6fd7706464d5961e45c1bf7cad5e9c7f67a1