[debugger] Expressions for mem commands.
Allows parsing expressions for the "mem-read" ("x") and "mem-analyize"
commands. This allows things like:

  [zxdb] x &foo

which is useful and people are accustomed to typing in GDB. This can
also automatically compute the size of the object being dumped when
given a typed pointer, which makes the output more useful.

Some helper functions are added for evaluating expressions in the
context of commands, and to optionally follow references (whan one wants
in this context should there be a reference to a pointer).

Previously all expression evaluation needed to be done in the context of
a stack frame. This is not desirable for the "mem-read" command which
should be able to read process memory even when the current thread is
running.

To solve this the FrameSymbolDataProvider is split apart into three.
Existing code can still use the existing FrameSymbolDataProvider. Code
that only has a process and no frame can now read memory but not
registers via the ProcessSymbolDataProvider, and without a process one
can use the SymbolDataProvider directly which allows the debugger to be
used like a calculator when there is no process running.

Change-Id: I53724c85c65950cf179208b8da0b92e00fc6b9b4