[arm64][exceptions][debugger] Integrated debugger HW exceptions.
arm64 was missing a key piece of integration for hardware exceptions: a
way to differentiate between a single step and a hw exception
(breakpoint). Before this CL it always assumed a hw exception is a
single step.

In order to differentiate, the kernel exposes the ESR register which
lets the debugger determine what was the last exception triggered. In
general this is not needed because the kernel already types the
exception, but ZX_HW_BREAKPOINT_EXCEPTION is ambiguous so it needs more
information. The same thing is done for x64 with DR6.

TEST=- Manual
     - BreakpointIntegration test:
src/developer/debug/debug_agent/integration_tests/breakpoint_test.cc.

NOTE: This test is currently disabled as we need to specify that it must
not run on QEMU.

Change-Id: I4800987ce203fcbe6cc75f31afb6c286bc6e27ed