[pseudo-fs] move most of handle_request to method on connection
for the most part, handle_request operates the same between different
types of pseudo files. the pseudo file implementations get the first
pass at the request stream, to filter out any behaviors on which they
differ, and to handle Close and Clone, which can't be done in the
connection object.

Change-Id: I01cdb4c7c8ed4fdb40be3221aa7bf067b1b23ba7