[appmgr] add isolated-cache-storage feature
Add the "isolated-cache-storage" feature, which causes the "/cache" path
in a component's namespace to be populated with an isolated path into
minfs. This functions identically to "isolated-persistent-storage", but
is backed by "/data/cache" instead of "/data", and the contents of
"/data/cache" will be deleted in the event of storage pressure.

CF-699 #done

Change-Id: Ic6aad3be8fe868ef8bca847ea1975f7dd65f932f