[zircon][dev][goldfish] Add goldfish pipe driver
This adds an initial goldfish pipe driver for zircon.
Goldfish pipe is a virtual device that intends to provide
a fast communication channel between the guest system and
the QEMU emulator.

Usage from the guest is simply as follows:

int fd = open("/dev/class/goldfish-pipe/000", O_RDWR);
... write() or read() through the pipe.

This driver doesn't deal with the exact protocol used. After
opening a pipe, write a null-terminated string to indicate
which service to establish a connection to:

const char*  msg = "<pipename>";
if (write(fd, msg, strlen(msg) + 1) < 0) {
  ... could not connect to <pipename> service
  close(fd);
}

After this, simply read() and write() to communicate with the
service.

The current implementation of this driver performs unnecessary
copy operations when sending and receiving data. This can be
improved in future versions of the driver by exposing a FIDL
interface that allows clients to use VMOs for data transfers.

Test: /boot/test/sys/goldfish-test
Change-Id: I3ac62aa8c34d17eb1ffda5f5cd4a8df31d8f1523