[pseudo-fs] watchers: Multiple names and AsRef<str>
While not used in the current code, lazy directories may triger add or
remove events that include multiple entry names.  It makes sense to pack
as much names as possible in the same message, instead of sending each
one individually.

Generalization for `AsRef<str>` is also useful for the lazy directory
case, where the entry names are not statically available and we do not
want to generate a list of all the entry names at any point in time.
See `Watchers::add()` comment for additional details.

Change-Id: I805ee520a0c36868ee6bbcf2a833e9e30d25d3a5