[go] remove zx.Error by-value assertions
This is the fourth step in transitioning the error implementation on
zx.Error from the value type to the pointer. It is safer to implement
interfaces on pointer types for the reason evident in this chnage; an
interface implemented by a value type is also implicitly implemented by
that value's pointer type, which requires type assertions to try both
cases. On the other hand, interface implementations by the pointer type
do not imply an implementation by the type's value type.

Change-Id: I009b7c9f9e12e4fbcd7f5dac38e88913939ca349