[media] remove lib/media/timeline dependency on fxl and fsl
This CL changes the garnet/public/lib/media/timeline library so it no
longer depends on libfxl or libfsl. This change is preliminary to
moving garnet/public/lib/media/timeline into sdk, where fxl and fsl
dependencies are forbidden. timeline.h is removed altogether, because
there are common replacements for everything it provides and because
it deals heavily in fxl::TimeDelta.

TEST: no behavior change

Change-Id: Iba9d4f5f088194ba792ab0a3d815ae624e6af406