[test spec] Introduce test resources
Valid only for linux/mac tests, this abstracts the metadata dance that
needs to take place in order for the infrastructure to respect the
runtime dependencies that test take (i.e., so the former will be
available on the bots when the test is executed).

Resources dictate where in source or the build directory files may be
placed, relative to the test executable.

Note: this change moves the test director of dart_test from
$root_out_dir to $target_gen_dir, as relative paths in the test are
treated as relative to the latter.

Change-Id: I8be94f859c400875801ebadbdd83c9fc9122c1f1