[media] remove lib/media/timeline dependency on fxl and fsl
This CL changes the garnet/public/lib/media/timeline library so it no
longer depends on libfxl or libfsl. This change is preliminary to
moving garnet/public/lib/media/timeline into sdk, where fxl and fsl
dependencies are forbidden. timeline.h is removed altogether, because
there are common replacements for everything it provides and because
it deals heavily in fxl::TimeDelta.

TEST: no behavior change

Change-Id: Ib216556198c832eea43df56c568f7a8d4d56f4f0