[debugger] Send configurations over to the agent.
This will permit to modify some behaviour on the fly for the agent.
Much of the code is IPC boilerplate.

Implemented the quit on exit command, which now doesn't depend on the
client. The current way it was implemented has a lot of issues with
correct timings of when is it called, because of destructor timings.

With this change, even is the client crashes, we won't be leaking the
agent.

TEST= - Unit test for protocol message
      - Manual by quitting the client.
      - Manual by killing the client.

Change-Id: Ia56a155db25d3528ddb1399a3d30738471710450