[fuchsia-component] Fix hang when `cd` or `find`
Sometimes the `find` command would open a service node so that later
it can obtain more metadata on that node. As for `cd`, it would open
a node with the expectation that the filesystem server rejects the
request if the node is not a directory. In those cases, because
ServiceFs automatically connect to the FIDL service at the specified
path, it causes those commands to hang.

This commit fixes these issues by handling Open request differently for
service node when OPEN_FLAG_DIRECTORY, OPEN_FLAG_NODE_REFERENCE, or
OPEN_FLAG_DESCRIBE is passed in. In those cases, we no longer connect to
the FIDL service.

DNO-460 #comment

Test:
1. Verify that `find /hub -name blah` doesn't hang
2. Verify that `cd` onto a FIDL service hosted by wlanstack2 doesn't
   hang
3. Sanity test wlanstack2:
   - Connect with `wlan` and `wlantool2` still work
   - Unplugging and replugging wireless dongle
   - Running `iquery` to retrieve Inspect dump

Change-Id: I46bddd29c2c365ba6563a1942a3c6c96eb81fb7d