[text] Split massive ime.rs file into three smaller files
Optimally, we'd reduce how much these files reach into each others state
as well, and reduce the need to have items on ImeState be `pub`. But
figured that was best to put in another, separate change, so that the
diff would better highlight that change, instead of being hidden in a
massive entire-file diff.

Change-Id: I99284ac444ca1e7333d1bc982df72835abaa9d36