[debugger] Enhance name finding.
Adds support for finding function names. This is not currently hooked up
to anything. In the future this will be used for completion of
breakpoint locations and could also be used for supporting taking
function addresses in expressions.

Adds support for prefix searching names and returning multiple results
up to some maximum.

Adds support for customizing the types of results to find. This allows,
for example, only finding function names and but not variable names
(this will be used for breakpoint location completion in a future
patch since we don't want to complete many types).

Update the IndexWalker to handle walking into identifiers that have a
"::" which indicates the global namespace.

Change-Id: I806832a130c9884acd97d2ffbf894b406203ad87