[wlan][driver] Don't add wlanif devices
The wlanif device becomes obsolete once a driver supports the new
SME channel approach as the device is only used to get a channel
via IOCTLs to the MLME.
The change is guarded by a feature flag such that only supported
drivers skip adding a wlanif device.
Once all SoftMAC and FullMAC drivers support the new approach,
the wlanif device can be removed entirely (Note: the wlanif-impl
device and its driver are still required; in fact there are plans
to use these for SoftMAC devices as well).

WLAN-927 #comment

Test=Sanity check broadcom and ath10k

Change-Id: If04b3e19c09562005cec70b4fb6c63134e5cbdf7