[zxcrypt] Support implicit device keys, use in boot/paver
We build upon the new kms-stateless library which abstracts over access
to hardware-unique keys.

New methods in FdioVolume and FdioVolumeManager allow creating,
unlocking, and unsealing the device with an implementation-defined key.

For now, in practice, this will continue using only the null key until
we change the `zxcrypt_key_source` build argument, which still reads
"null" for all products.

Related bugs: ZX-2670, ZX-3367

Tests: I've successfully tested this end-to-end with some additional
changes not present in this branch that will take some time to make
their way into appropriate repos:

* BL32 changes to enable reading efuses
* a self-signed copy of the keysafe TA, which we'll need to get into CIPD
* a patch to declare the keysafe TA as firmware for the optee driver
* a patch to set `zxcrypt_key_source` to `tee` for astro

We can land this patch and its parent safely and see no observable
change in behavior for now.  We cannot safely flip the build argument
until we've sorted out the bootloader/firmware situation, and that's
something we need to coordinate with other folks.

This general codepath will more broadly be tested as an integration test
once IN-302 is resolved.

Change-Id: I6b53f18f0876f6ebde75db0b2ff543dbc763946c