[modular][testing] Implement fuchsia.modular.testing.TestHarness
Details:
* The services from TestHarness.GetService() are supplied using a
session agent which supplied to the modular config. This session agent
URL is intercepted and implemented by
TestHarnessImpl::InterceptedSessionAgent. GetService() requests are
buffered until this session agent starts up.

* InterceptedComponentImpl implements and manages the fidl bindings for
the |InterceptedComponent| protocol. It signals TestHarnessImpl to
remove the InterceptedComponentImpl in the event that the interface closes.

MF-244
MF-249

Test: TestHarnessImpl.* tests shell and component (module) interception,
and default modular configuration.

Change-Id: I84f5f845490cdeb8bd892a35fe8c9e46e9f05fe1