[framebuffer] Remove page-flipping support from libframebuffer
Since the mali driver no longer uses libframebuffer, there's no consumer
of its page-flipping support. The current API isn't a great fit for
sysmem (e.g. it assumes the stride is fixed despite what the client
might want, and that independent vmos can be allocated), so remove the
page-flipping support for now.

We can add a modifed version of this API back if we get a new consumer
of it.

Change-Id: I40c07879b1104df549255e3c1884adebda246659