[fidlc][c] Generate proper struct defs for xunion and tables
Because we do not support xunions and tables in the C bindings, we need
to refer to the raw FIDL type (fidl_xunion_t and fidl_table_t
respectively), and leave it up to users to figure out the rest. Before,
we incorrectly generated a ref, which pointed to nothing (since it does
not get generated).

Also adding fidl_table_t, which we hadn't needed before. (We could now
use it in the walker to make things a tiny little bit more typesafe.)

FIDL-578 #done

Change-Id: Id563cf8d5ace42342fabb613ff3f7e3dd275523c