[scudo] Various code generation optimizations
Scudo has to not be compiled with `-ffreestanding` so that various small
fixed size `memcpy()` can be optimized away, as such we can't depend on
musl_internal.

On the other hand, the compiler is pretty aggressive at inlining things
resulting in some code bloat, where a large function is duplicated many
times. So we add the `ALWAYS_INLINE` or `NOINLINE` attributes to various
functions to make the generated code less clunky.

Finally, include some changes from upstream reviews.

Change-Id: Ie80d51e84b7d496e09934e1a59b1ea70eedf710e