[pkg] RepositoryConfig should use RepoUri and FuchsiaPkgUri
This updates fidl_fuchsia_pkg_ext::RepositoryConfig to treat
the repo_url as a RepoUri and update_package_uri to use
the FuchsiaPkgUri type. This allows us to assert at compile time
that these fields have the proper type, and saves effort having
to translate them to RepoUri/FuchsiaPkgUri later.

Change-Id: I0bfd2d892b8869f2e436dbba9eb356af31cf652c