[netsvc][bootserver] Send a sentinel file with board name
If board name is specified via flag to flag input to bootserver, it will
first send a file with the board name inside to the device. On the
device, netsvc will handle this file in a special manner, and fail the
transfer if the board name inside doesn't match the board name of the
device itself.

This change prevents eroneous build images from being paved. The next
change will update the build system to start sending the board name
during invocations from fx pave.

ZX-3794

Change-Id: I1c0bcb1503e06f6773896be929665b382f1100a3