[debugger] Delegate initial thread control to client.
Previously both the debug_agent and the client had code to decide when a
new thread should be resumed. In cases where the thread should be
resumed, this meant the logic was duplicated and the thread was resumed
twice.

This is mostly benign but if a thread hits a breakpoint eatly in its
execution, the second resume from the client can accidentally resume
from the exception.

This change modes all decisions about new thread running state to the
client. Tests are updated accordingly.

Change-Id: Id4bb813ab2f2511ad3674737d9757b8b97a05799