[debugger] Ensure file names are unique.
In an effort to keep output succinct, source code locations would only
show the final name part of a file path. But sometimes this is not
enough information to unqiuely identify a source file in a process.

Enhances file name printing to compute the shortest unique name for a
file in the current process. This still keeps names short but now
they're not ambiguous.

Change-Id: Ib1ade7c8f55b001bd93e3265649dced173f9193f